move_as_list_to_common
